Refactor mrow to common and chtml versions
