+++ mathjax3-ts/output/chtml/Wrappers/mrow.ts
@@ -21,12 +21,9 @@
  * @author dpvc@mathjax.org (Davide Cervone)
  */
 
-import {CHTMLWrapper} from '../Wrapper.js';
-import {CHTMLWrapperFactory} from '../WrapperFactory.js';
-import {MmlMrow, MmlInferredMrow} from '../../../core/MmlTree/MmlNodes/mrow.js';
-import {MmlNode} from '../../../core/MmlTree/MmlNode.js';
-import {BBox} from '../BBox.js';
-import {DIRECTION} from '../FontData.js';
+import {CHTMLWrapper, CHTMLConstructor} from '../Wrapper.js';
+import {CommonMrow, CommonMrowInterface, MrowConstructor} from '../../common/Wrappers/mrow.js';
+import {CommonInferredMrow, CommonInferredMrowInterface} from '../../common/Wrappers/mrow.js';
 
 /*****************************************************************/
 /**
@@ -36,24 +33,7 @@ import {DIRECTION} from '../FontData.js';
  * @template T  The Text node class
  * @template D  The Document class
  */
-export class CHTMLmrow<N, T, D> extends CHTMLWrapper<N, T, D> {
-    public static kind = MmlMrow.prototype.kind;
-
-    /**
-     * @override
-     * @constructor
-     */
-    constructor(factory: CHTMLWrapperFactory<N, T, D>, node: MmlNode, parent: CHTMLWrapper<N, T, D> = null) {
-        super(factory, node, parent);
-        this.stretchChildren();
-        for (const child of this.childNodes) {
-            if (child.bbox.pwidth) {
-                this.bbox.pwidth = BBox.fullWidth;
-                break;
-            }
-        }
-    }
-
+export class CHTMLmrow<N, T, D> extends CommonMrow<N, T, D, CHTMLConstructor<N, T, D>>(CHTMLWrapper) {
     /**
      * @override
      */
@@ -78,65 +58,15 @@ export class CHTMLmrow<N, T, D> extends CHTMLWrapper<N, T, D> {
         }
     }
 
-    /**
-     * Handle vertical stretching of children to match height of
-     *  other nodes in the row.
-     */
-    public stretchChildren() {
-        let stretchy: CHTMLWrapper<N, T, D>[] = [];
-        //
-        //  Locate and count the stretchy children
-        //
-        for (const child of this.childNodes) {
-            if (child.canStretch(DIRECTION.Vertical)) {
-                stretchy.push(child);
-            }
-        }
-        let count = stretchy.length;
-        let nodeCount = this.childNodes.length;
-        if (count && nodeCount > 1) {
-            let H = 0, D = 0;
-            //
-            //  If all the children are stretchy, find the largest one,
-            //  otherwise, find the height and depth of the non-stretchy
-            //  children.
-            //
-            let all = (count > 1 && count === nodeCount);
-            for (const child of this.childNodes) {
-                const noStretch = (child.stretch.dir === DIRECTION.None);
-                if (all || noStretch) {
-                    const {h, d} = child.getBBox(noStretch);
-                    if (h > H) H = h;
-                    if (d > D) D = d;
-                }
-            }
-            //
-            //  Stretch the stretchable children
-            //
-            for (const child of stretchy) {
-                child.coreMO().getStretchedVariant([H, D]);
-            }
-        }
-    }
-
 }
 
 /*****************************************************************/
 /**
  *  The CHTMLinferredMrow wrapper for the MmlInferredMrow object
+ *
+ * @template N  The HTMLElement node class
+ * @template T  The Text node class
+ * @template D  The Document class
  */
-
-export class CHTMLinferredMrow<N, T, D> extends CHTMLmrow<N, T, D> {
-    public static kind = MmlInferredMrow.prototype.kind;
-
-    /**
-     * Since inferred rows don't produce a container span, we can't
-     * set a font-size for it, so we inherit the parent scale
-     *
-     * @override
-     */
-    protected getScale() {
-        this.bbox.scale = this.parent.bbox.scale;
-        this.bbox.rscale = 1;
-    }
+export class CHTMLinferredMrow<N, T, D> extends CommonInferredMrow<N, T, D, MrowConstructor>(CHTMLmrow) {
 }
